requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.28.0
discord.py>=2.3.0
openpyxl>=3.1.0
playwright>=1.40.0
